(libT) #5267: 'speed limits wrap around too low number (~8589 kB/s)': patch by const

+++ libtransmission/bandwidth.c
@@ -184,7 +184,7 @@ allocateBandwidth (tr_bandwidth  * b,
   if (b->band[dir].isLimited)
     {
       const uint64_t nextPulseSpeed = b->band[dir].desiredSpeed_Bps;
-      b->band[dir].bytesLeft = (unsigned int)(nextPulseSpeed * period_msec) / 1000u;
+      b->band[dir].bytesLeft = nextPulseSpeed * period_msec / 1000u;
     }
 
   /* add this bandwidth's peer, if any, to the peer pool */
